backend/requirements.txt
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
lazr.restfulclient==0.14.6
pillow==12.1.1
python-dotenv==1.2.1
